user_api/pyproject.toml
[project]
name = "user-api"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "asgiref>=3.9.1",
    "fastapi>=0.116.1",
    "grpcio>=1.74.0",
    "grpcio-tools>=1.74.0",
    "loguru>=0.7.3",
    "oss2>=2.19.1",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "redis[hiredis]>=6.4.0",
    "unisms>=0.0.7",
    "uvicorn>=0.35.0",
]
